모닝 커밋

+++ src/net/mbiz/library/ui/library/myPage/MyPageTablePanel.java
@@ -3,7 +3,6 @@ package net.mbiz.library.ui.library.myPage;
 import java.awt.BorderLayout;
 import java.awt.Color;
 import java.awt.Dimension;
-import java.awt.Image;
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
 import java.awt.event.MouseEvent;
@@ -12,11 +11,13 @@ import java.text.SimpleDateFormat;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.Date;
+import java.util.List;
 
 import javax.swing.BorderFactory;
 import javax.swing.Box;
-import javax.swing.ImageIcon;
+import javax.swing.DefaultComboBoxModel;
 import javax.swing.JButton;
+import javax.swing.JComboBox;
 import javax.swing.JLabel;
 import javax.swing.JOptionPane;
 import javax.swing.JPanel;
@@ -29,32 +30,38 @@ import net.mbiz.library.data.AddBookList;
 import net.mbiz.library.data.AddBorrowList;
 import net.mbiz.library.data.BookVO;
 import net.mbiz.library.data.BorrowVO;
+import net.mbiz.library.main.LibraryMain;
 import net.mbiz.library.ui.common.CommonConstants;
-import net.mbiz.library.ui.main.LibraryMain;
 
 public class MyPageTablePanel extends JPanel implements ActionListener, MouseListener{
 	
 	private JPanel pnBody;
+	//body
 	private JPanel pnTbl;
 	private JPanel pnHeader;
 	private JPanel pnFooter;
-	private JPanel pnSch;
-	private JPanel pnPadding;
+	//header
 	private JPanel pnTitle;
-	private JPanel pnPvs;
-	private JPanel pnBtnSet;
+	private JPanel pnEast;
+	private JPanel pnWest;
+
+	private JPanel pnCbb;
 	
+	private JPanel pnBtnSet;
 	private JLabel title;
 	private JLabel pvsLbl;
-	
-	private JTextField schFd;
-	
+
 	private JButton schBtn;
 	private JButton pvsBtn;
 	private JButton returnBtn;
 	private JButton deleteBtn;
 	
+	private JTextField schFd;
+	
 	private JTable borrowTbl;
+	
+	private JComboBox<String> cbbSearch;
+	private List<BorrowVO> checkedList = new ArrayList<>(); 
 	
 	public MyPageTablePanel(LibraryMain f) {
 		jbInit();
@@ -78,7 +85,6 @@ public class MyPageTablePanel extends JPanel implements ActionListener, MouseLis
 		this.pnBody = new JPanel();
 		pnBody.setLayout(new BorderLayout());
 		pnBody.setBackground(Color.DARK_GRAY);
-
 		
 		//NORTH
 		this.pnHeader = new JPanel();
@@ -87,20 +93,17 @@ public class MyPageTablePanel extends JPanel implements ActionListener, MouseLis
 		pnHeader.setBackground(CommonConstants.COLOR_CONTENT_BACKGROUND);
 		pnHeader.add(Box.createHorizontalStrut(920), BorderLayout.WEST);
 		pnHeader.add(Box.createVerticalStrut(15), BorderLayout.SOUTH);
-
-		
-		//CENTER - 테이블 영역 패널
+		//CENTER
 		this.pnTbl = new JPanel(); 
 		pnTbl.setLayout(new BorderLayout());
 		pnTbl.setBackground(CommonConstants.COLOR_CONTENT_BACKGROUND);
 		this.borrowTbl  = new JTable(CommonConstants.bwModel); // 데이터가 들어가는 테이블 
 		borrowTbl.setRowHeight(32);
 		borrowTbl.setFont(CommonConstants.FONT_BASE_17);
-		// 테이블 스크롤
+
 		JScrollPane sclpn = new JScrollPane(borrowTbl);
 		sclpn.setCorner(JScrollPane.UPPER_LEFT_CORNER, borrowTbl.getTableHeader());
-		
-		
+		pnTbl.add(sclpn, BorderLayout.CENTER);
 		//SOUTH		
 		this.pnFooter = new JPanel();
 		pnFooter.setLayout(new BorderLayout());
@@ -110,39 +113,73 @@ public class MyPageTablePanel extends JPanel implements ActionListener, MouseLis
 		
 		
 		
+		
+		
 		/* pnHeander - pnTitle(NORTH), pnSch(EAST), pnPvs(WEST) */
 		// NORTH
 		this.pnTitle = new JPanel();
 		pnTitle.setLayout(new BorderLayout());
 		pnTitle.setPreferredSize(new Dimension(0,55));
 		pnTitle.setBackground(CommonConstants.COLOR_CONTENT_BACKGROUND);
+		
 		this.title = new JLabel("나의 대출리스트");
 		title.setFont(CommonConstants.FONT_TITLE_25);
 		title.setHorizontalAlignment(JLabel.LEFT);
 		title.setBorder(BorderFactory.createEmptyBorder(30, 30, 30, 30));
+		
+		pnTitle.add(title, BorderLayout.WEST);
 
 		//EAST
-		this.pnSch = new JPanel();
-		pnSch.setLayout(new BorderLayout());
-		pnSch.setBackground(CommonConstants.COLOR_CONTENT_BACKGROUND);
-		pnSch.setBorder(BorderFactory.createEmptyBorder(33, 0, 0, 0));
+		this.pnEast = new JPanel();
+		pnEast.setLayout(new BorderLayout());
+		pnEast.setBackground(CommonConstants.COLOR_CONTENT_BACKGROUND);
+		pnEast.setBorder(BorderFactory.createEmptyBorder(33, 0, 0, 0));
 		
 		//WEST
-		this.pnPvs = new JPanel();
-		pnPvs.setLayout(new BorderLayout());
-		pnPvs.setBorder(BorderFactory.createEmptyBorder(33, 0, 0, 0));
-		pnPvs.setBackground(CommonConstants.COLOR_CONTENT_BACKGROUND);
-		/* 전체 리스트로. */
+		this.pnWest = new JPanel();
+		pnWest.setLayout(new BorderLayout());
+		pnWest.setBorder(BorderFactory.createEmptyBorder(33, 0, 0, 0));
+		pnWest.setBackground(CommonConstants.COLOR_CONTENT_BACKGROUND);
+
 		this.pvsBtn = new JButton();
 		pvsBtn.setPreferredSize(new Dimension(200, 30));
 		pvsBtn.setBorderPainted(false);
 		pvsBtn.setBackground(CommonConstants.COLOR_CONTENT_BACKGROUND);
-
 		this.pvsLbl = new JLabel("전체보기");
 		pvsLbl.setFont(CommonConstants.FONT_BASE_17);
 		pvsLbl.setHorizontalAlignment(JLabel.CENTER);
+	
 		pvsBtn.add(pvsLbl);	
+		pnWest.add(pvsBtn, BorderLayout.WEST);
 		
+		
+		
+		/* pnEast - pnCbb(WEST), schFd(CENTER), schBtn(EAST) */
+		this.schFd = new JTextField();
+		schFd.setPreferredSize(new Dimension(550, 30));
+		schFd.setFont(CommonConstants.FONT_BASE_17);
+		
+		this.schBtn = new JButton("검색");
+		schBtn.setPreferredSize(new Dimension(70, 30));
+		schBtn.setFont(CommonConstants.FONT_BASE_12);
+
+		this.pnCbb = new JPanel();
+		pnCbb.setBackground(CommonConstants.COLOR_CONTENT_BACKGROUND);
+
+		this.cbbSearch = new JComboBox<>();
+		cbbSearch.setModel(new DefaultComboBoxModel<>(new String[] {"도서명","저자"}));
+		cbbSearch.setFont(CommonConstants.FONT_BASE_15);
+		cbbSearch.setPreferredSize(new Dimension(90,50));
+
+		pnCbb.add(cbbSearch);
+		pnEast.add(schFd, BorderLayout.CENTER);
+		pnEast.add(pnCbb, BorderLayout.WEST);
+		pnEast.add(schBtn, BorderLayout.EAST);
+		
+		
+		pnHeader.add(pnTitle, BorderLayout.NORTH);
+		pnHeader.add(pnEast, BorderLayout.EAST);
+		pnHeader.add(pnWest, BorderLayout.WEST);
 		
 		
 		
@@ -153,7 +190,6 @@ public class MyPageTablePanel extends JPanel implements ActionListener, MouseLis
 		pnBtnSet.add(Box.createVerticalStrut(10), BorderLayout.NORTH);
 		pnBtnSet.add(Box.createHorizontalStrut(10), BorderLayout.CENTER);
 		
-		
 		this.returnBtn = new JButton("반납");
 		returnBtn.setPreferredSize(new Dimension(70, 70));
 		returnBtn.setFont(CommonConstants.FONT_BASE_15);
@@ -161,49 +197,16 @@ public class MyPageTablePanel extends JPanel implements ActionListener, MouseLis
 		deleteBtn.setPreferredSize(new Dimension(70, 70));
 		deleteBtn.setFont(CommonConstants.FONT_BASE_15);
 		
-		
-		
-		/* pnSch - schFd(WEST), pnPadding(CENTER), schBtn(EAST) */
-		// WEST - 검색어 입력란
-		this.schFd = new JTextField();
-		schFd.setPreferredSize(new Dimension(550, 30));
-		schFd.setFont(CommonConstants.FONT_BASE_17);
-		
-		// EAST - 검색 버튼
-		this.schBtn = new JButton();
-		schBtn.setPreferredSize(new Dimension(70, 30));
-		schBtn.setFont(CommonConstants.FONT_BASE_12);
-
-		// CENTER
-		this.pnPadding = new JPanel();
-		pnPadding.setPreferredSize(new Dimension(10, 10));
-		pnPadding.setBackground(CommonConstants.COLOR_CONTENT_BACKGROUND);
-
-
-		//footer
 		pnBtnSet.add(returnBtn, BorderLayout.WEST);
 		pnBtnSet.add(deleteBtn, BorderLayout.EAST);
 		pnFooter.add(pnBtnSet, BorderLayout.EAST);
 		
-		//header
-		pnSch.add(pnPadding, BorderLayout.CENTER);
-		pnSch.add(schFd, BorderLayout.WEST);
-		pnSch.add(schBtn, BorderLayout.EAST);
 		
-		pnPvs.add(pvsBtn, BorderLayout.WEST);
-		pnTitle.add(title, BorderLayout.WEST);
-		
-		pnHeader.add(pnTitle, BorderLayout.NORTH);
-		pnHeader.add(pnSch, BorderLayout.EAST);
-		pnHeader.add(pnPvs, BorderLayout.WEST);
 
-		//center
-		pnTbl.add(sclpn, BorderLayout.CENTER);
 		
 		pnBody.add(pnFooter, BorderLayout.SOUTH);
 		pnBody.add(pnTbl, BorderLayout.CENTER);
 		pnBody.add(pnHeader, BorderLayout.NORTH);
-		
 		this.add(pnBody, BorderLayout.CENTER);
 		
 		
@@ -220,13 +223,10 @@ public class MyPageTablePanel extends JPanel implements ActionListener, MouseLis
 	 * 대출 기록 테이블을 출력하는 메서드.
 	 * */
 	private void initTable() {
-		SimpleDateFormat sdf = new SimpleDateFormat("yyyy.MM.dd");
-		String topHeader[] = {"no", "도서명", "저자","대출일", "반납예정일", "반납일", "연체일", "반납", "삭제"};	
-		int col[] = {60 ,630 ,290 ,173 ,173 ,173, 90, 80, 80};
+		String topHeader[] = {"no", "도서명", "저자", "대출일", "반납예정일", "반납일", "연체일", "check"};	
+		int[] col = {60, 60, 600, 296, 200, 180, 180, 180 };
 		
 		CommonConstants.bwModel = new BeanTableModel<BorrowVO>(topHeader, col) {
-			
-			// 객체의 컬럼 별 데이터를 한꺼번에 테이블에 뿌려줍니다.
 			@Override
 			public Object getValueByColumIndex(int row, int col) {
 				BorrowVO bv = getRowAt(row);
@@ -238,22 +238,22 @@ public class MyPageTablePanel extends JPanel implements ActionListener, MouseLis
 				
 			}
 		};
-		
 		CommonConstants.setTableModelColumnWithCommonTableRenderer(borrowTbl, CommonConstants.bwModel);
-
 		
-		
-		
+		//EVENT
 		borrowTbl.addMouseListener(this);
 	
 	}
 	
+	
+	
+	//---------------------------ActionListener, MouseListener------------------------------------	
 	@Override
 	public void actionPerformed(ActionEvent e) {
 		if (e.getSource().equals(pvsBtn)) {
 			CommonConstants.repaintBorrowTable();
 		} else if (e.getSource().equals(schBtn)) {
-			searchBorrow();
+			getSearchBorrowList();
 		} else if (e.getSource().equals(returnBtn)) {
 			returnBook();
 		} else if (e.getSource().equals(deleteBtn)) {
@@ -320,16 +320,35 @@ public class MyPageTablePanel extends JPanel implements ActionListener, MouseLis
 	 * 대출 기록 삭제 시, 확인 메세지를 띄워주는 메서드.
 	 */
 	private void getDeleteMessege() {
-		BorrowVO seletedVO = AddBorrowList.borrowList.get(borrowTbl.getSelectedRow());
-
-		int rslt = JOptionPane.showConfirmDialog(null, " '" + seletedVO.getBookNm()+ "' " + " 대출 기록을 삭제 하시겠습니까?", seletedVO.getBookNm(), JOptionPane.YES_NO_OPTION);
-		if (rslt == JOptionPane.YES_OPTION) { // '예' 선택
-			deleteBorrowVO(seletedVO);
-			JOptionPane.showMessageDialog(null,"대출 기록이 삭제 되었습니다.");
-		} else { 
-			System.out.println(seletedVO.getBookNm() + " 대출 기록 삭제를 취소합니다.");
-			JOptionPane.showMessageDialog(null, "삭제가 취소 되었습니다.", seletedVO.getBookNm(), JOptionPane.INFORMATION_MESSAGE);
+		
+		if (checkedList.size() > 0) {
+			int rslt = JOptionPane.showConfirmDialog(null,  checkedList.size() + "개의" + " 대출 기록을 삭제 하시겠습니까?", "대출 기록을 삭제합니다.", JOptionPane.YES_NO_OPTION);
+			if (rslt == JOptionPane.YES_OPTION) { // '예' 선택
+				deleteCheckedList(checkedList);
+				JOptionPane.showMessageDialog(null,"삭제 완료 되었습니다.");
+			} else { 
+				JOptionPane.showMessageDialog(null, "삭제 취소 되었습니다.");
+			}	
+			
+		} else  {
+			if (borrowTbl.getSelectedRow()> -1) { // row가 선택되어 있을 때
+				BorrowVO seletedVO = AddBorrowList.borrowList.get(borrowTbl.getSelectedRow());
+				
+				int rslt = JOptionPane.showConfirmDialog(null, " '" + seletedVO.getBookNm()+ "' " + " 대출 기록을 삭제 하시겠습니까?", seletedVO.getBookNm(), JOptionPane.YES_NO_OPTION);
+				if (rslt == JOptionPane.YES_OPTION) { // '예' 선택
+					deleteBorrowVO(seletedVO);
+					JOptionPane.showMessageDialog(null,"대출 기록이 삭제 되었습니다.");
+				} else { 
+					System.out.println(seletedVO.getBookNm() + " 대출 기록 삭제를 취소합니다.");
+					JOptionPane.showMessageDialog(null, "삭제가 취소 되었습니다.", seletedVO.getBookNm(), JOptionPane.INFORMATION_MESSAGE);
+				}
+			} else {
+				JOptionPane.showMessageDialog(null, "삭제할 기록을 선택하세요.","선택된 행 없음", JOptionPane.INFORMATION_MESSAGE);
+			}
 		}
+		
+		
+
 	}
 
 	/**
@@ -377,20 +396,28 @@ public class MyPageTablePanel extends JPanel implements ActionListener, MouseLis
 	/**
 	 * 대출 기록을 검색하는 메서드
 	 */
-	private void searchBorrow() {
+	private void getSearchBorrowList() {
 		
 		if (!schFd.getText().isEmpty() && !schFd.getText().equals("")) {
 			CommonConstants.bwModel.removeAll();
+			
 			for (BorrowVO bv : AddBorrowList.borrowList) {
-				if (bv.getBookNm().contains(schFd.getText())) {
-					CommonConstants.bwModel.addData(bv);
-					System.err.println("여기는 searchBorrow. for문 안. 검색결과 bv는? --->" + bv);
+				String cbb = (String) cbbSearch.getSelectedItem();
+				
+				if (cbb.equals("도사명")) {
+					if (bv.getBookNm().contains(schFd.getText())) {
+						CommonConstants.bwModel.addData(bv);
+					}	
+				} else if (cbb.equals("저자")) {
+					if (bv.getBookWtr().contains(schFd.getText())) {
+						CommonConstants.bwModel.addData(bv);
+					}
 				}
 			}
 			this.borrowTbl.setModel(CommonConstants.bwModel);
 
 		} else {
-			JOptionPane.showMessageDialog(pnBody, "검색어를 입력해주세요.");
+			CommonConstants.repaintBorrowTable();
 		}
 	}
 	
@@ -403,6 +430,22 @@ public class MyPageTablePanel extends JPanel implements ActionListener, MouseLis
 		
 		CommonConstants.repaintBorrowTable();
 			
+	}
+	
+	
+	/**
+	 * 체크박스로 선택된 대출 데이터를 삭제하는 메서드.
+	 * @param vo BorrowVO
+	 */
+	private void deleteCheckedList(List<BorrowVO> checkedList) {
+		System.err.println("여기 마이페이지 deleteCheckedList :::: "+checkedList);
+		for (BorrowVO vo : checkedList) {
+			deleteBorrowVO(vo);
+		}
+		CommonConstants.repaintBookTable();
+		borrowTbl.removeAll();
+		borrowTbl.setModel(CommonConstants.bwModel);
+		
 	}
 
 }
